Data recording will not auto pause if SD card ejected
SD card ejections will be recorded in the CSV file, along with power loss detection. In case of accidental SD card ejection, device will continue capturing data once SD card inserted again. On accidental power loss, device will note it in the CSV, but will not continue until connected to the dock program. This is done to sync time as the nrf52840 does not have a real time clock.